Refactored to add username global context
