+++ src/Contexts/UserContexts.jsx
@@ -1,0 +1,15 @@
+import { createContext, useContext, useState } from "react";
+
+const UserContext = createContext();
+
+export function UserProvider({ children }) {
+  const [username, setUsername] = useState("tickle122");
+
+  return (
+    <UserContext.Provider value={{ username, setUsername }}>
+      {children}
+    </UserContext.Provider>
+  );
+}
+
+export const useUser = () => useContext(UserContext);
