+++ src/main.jsx
@@ -4,14 +4,17 @@ import "./index.css";
 import "bootstrap/dist/css/bootstrap.min.css";
 import App from "./App.jsx";
 import { BrowserRouter } from "react-router-dom";
-import { ErrorProvider } from "./Contexts/Error.jsx";
+import { ErrorProvider } from "./Contexts/ErrorContext.jsx";
+import { UserProvider } from "./Contexts/UserContexts.jsx";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <BrowserRouter>
     <ErrorProvider>
-      <StrictMode>
-        <App />
-      </StrictMode>
+      <UserProvider>
+        <StrictMode>
+          <App />
+        </StrictMode>
+      </UserProvider>
     </ErrorProvider>
   </BrowserRouter>
 );
